+++ jquery.stl/jquery.stl.queue.js
@@ -1,5 +1,5 @@
 /**
- * jQuery STL stack Plugin v1.0b
+ * jQuery STL queue Plugin v1.0b
  * https://github.com/siciarek/jquery-stl
  *
  * Copyright 2012, Jacek Siciarek
@@ -9,14 +9,23 @@
  */
 (function ($, document, undefined) {
 
-    $.stl.stack = function (container) {
+    $.stl.queue = function (container) {
 
         var containerMembers = $.stl.getContainerAdaptorMembers({
-            front: true,
-            back: true
+            top: true
         });
 
-        var uniqueMembers = {};
+        var uniqueMembers = {
+            /**
+             * Remove element
+             * Removes the element on top of the stack, effectively reducing its size by one
+             *
+             * @return {*}
+             */
+            pop: function () {
+                return this.container.pop_front();
+            }
+        };
 
         var instance = $.extend({}, containerMembers, uniqueMembers);
         instance.new(container);