+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.2b
+---
+- Container classes: vector, deque, list
+- Container adaptors: stack, queue
+
 0.1b
 ---
 - Container classes: vector